cond: maybe fix cmpwi?

+++ src/instr/cond.c
@@ -11,11 +11,11 @@
 static void do_cmpi(struct _ppcemu_state *state, uint crfD, uint rA, u16 simm) {
 	u32 c, a = state->gpr[rA];
 
-	if (a < simm) /* TODO: sign-extend */
+	if ((i32)a < (i32)(i16)simm)
 		c = 0b1000;
-	else if (a > simm) /* TODO: sign-extend */
+	else if ((i32)a > (i32)(i16)simm)
 		c = 0b0100;
-	else
+	else /* equal */
 		c = 0b0010;
 
 	c |= !!(state->sprs[ppcemu_sprn_to_idx(PPCEMU_SPRN_XER)] & PPCEMU_XER_SO);
